+++ tests/testthat/test-oa_fetch.R
@@ -54,6 +54,26 @@ test_that("oa_fetch works", {
     cited_by_count = ">1000"
   )
   expect_s3_class(filtered_works, "data.frame")
+
+  expect_warning(oa_fetch(doi = "123"))
+})
+
+test_that("oa_fetch options$select works", {
+  skip_on_cran()
+
+  x <- oa_fetch(
+    entity = "works",
+    doi = c(
+      "10.1371/journal.pone.0266781",
+      "10.1371/journal.pone.0267149"
+    ),
+    options = list(select = c("doi", "id", "cited_by_count", "type")),
+    verbose = TRUE
+  )
+  expect_equal(
+    dim(x),
+    c(2, 4)
+  )
 })
 
 test_that("Error when input entity can't be matched", {
@@ -89,18 +109,17 @@ test_that("oa_fetch sample works", {
   random2021 <- oa_fetch(
     "works",
     publication_year = 2021,
-    sample = 20
+    options = list(sample = 20)
   )
   Sys.sleep(1/10)
   random10 <- oa_fetch(
     "works",
-    sample = 10,
-    seed = 1
+    options = list(sample = 10, seed = 1)
   )
 
   expect_equal(nrow(random2021), 20)
   expect_equal(nrow(random10), 10)
-  expect_error(oa_fetch("works", search = "open science", sample = 10))
+  expect_error(oa_fetch("works", search = "open science", options = list(sample = 10)))
 })
 
 
@@ -193,7 +212,6 @@ test_that("oa_fetch can combine (OR) more than 50 DOIs in a filter", {
   expect_equal(nrow(many_doi_results), length(valid_dois))
 })
 
-
 test_that("oa_fetch can combine (OR) more than 50 ORCIDs in a filter", {
   skip_on_cran()
 
@@ -267,3 +285,16 @@ test_that("oa_random works", {
   expect_equal(nrow(random_works), 1)
 })
 
+test_that("oa_fetch other entities works", {
+  skip_on_cran()
+
+  random_authors <- oa_fetch(entity = "authors", options = list(sample = 20))
+  random_venues <- oa_fetch(entity = "venues", options = list(sample = 20))
+  random_concepts <- oa_fetch(entity = "concepts", options = list(sample = 20))
+  random_institutions <- oa_fetch(entity = "institutions", options = list(sample = 20))
+
+  expect_equal(nrow(random_authors), 20)
+  expect_equal(nrow(random_venues), 20)
+  expect_equal(nrow(random_concepts), 20)
+  expect_equal(nrow(random_institutions), 20)
+})