+++ NEWS.md
@@ -1,4 +1,6 @@
 # openalexR (development version)
+* Arguments sample, seed, sort, and select are now grouped into `options`.
+* `abstract` now defaults to TRUE to avoid issues.
 * New argument to oa_fetch: api_key
 * Bug fixes
 * Improved documentation: group functions in Reference, details on search, etc.